Test with specific folder renaming to force update.
